madrox/onprem_cleanup

+++ components/automate-cli/cmd/chef-automate/cleanup.go
@@ -13,12 +13,6 @@ var cleanupFlags = struct {
 	onprem        bool
 }{}
 
-// const (
-// 	FSCRIPT = `systemctl stop chef-automate;
-// rm -rf /var/automate-ha;
-// rm -rf /hab`
-// )
-
 func init() {
 	RootCmd.AddCommand(cleanupCmd)
 	cleanupCmd.PersistentFlags().BoolVar(&cleanupFlags.onprem, "onprem", false, "Cleaning up all the instances related to onprem ")
@@ -46,18 +40,27 @@ func runCleanupCmd(cmd *cobra.Command, args []string) error {
 	writer.Printf(strings.Join(args, ""))
 	if isA2HARBFileExist() {
 		if cleanupFlags.onprem {
-			frontendIps := append(infra.Outputs.ChefServerPrivateIps.Value, infra.Outputs.AutomatePrivateIps.Value...)
-			backendIps := append(infra.Outputs.PostgresqlPrivateIps.Value, infra.Outputs.OpensearchPrivateIps.Value...)
+			FrontendIps := append(infra.Outputs.ChefServerPrivateIps.Value, infra.Outputs.AutomatePrivateIps.Value...)
+			BackendIps := append(infra.Outputs.PostgresqlPrivateIps.Value, infra.Outputs.OpensearchPrivateIps.Value...)
 			args = append(args, "--onprem")
-			// writer.Printf("IPs: " + strings.Join(frontendIps, "") + args[0])
-			for i := 0; i < len(frontendIps); i++ {
-				executeCleanupOnRemote(sshUser, sshPort, sskKeyFile, frontendIps[i], args[0])
+			for i := 0; i < len(FrontendIps); i++ {
+				executeCleanupOnRemote(sshUser, sshPort, sskKeyFile, FrontendIps[i], args[0])
 			}
-			for i := 0; i < len(backendIps); i++ {
-				executeCleanupOnBackendRemote(sshUser, sshPort, sskKeyFile, backendIps[i], args[0])
+			for i := 0; i < len(BackendIps); i++ {
+				executeCleanupOnRemote(sshUser, sshPort, sskKeyFile, BackendIps[i], args[0])
 			}
-
+			cleanUpScript := "hab pkg uninstall chef/automate-ha-deployment"
+			args := []string{
+				"-c",
+				cleanUpScript,
+			}
+			err := executeCommand("/bin/sh", args, "")
+			if err != nil {
+				return err
+			}
+			return nil	
 		}
+		
 	}
 	return nil
 }
@@ -73,9 +76,14 @@ func executeCleanupOnRemote(sshUser string, sshPort string, sshKeyFile string, i
 		writer.Errorf("Parsing key failed: %v", err)
 	}
 	config := &ssh.ClientConfig{
-		User:            sshUser,
-		Auth:            []ssh.AuthMethod{ssh.PublicKeys(signer)},
-		HostKeyCallback: ssh.InsecureIgnoreHostKey(),
+		User: sshUser,
+		Auth: []ssh.AuthMethod{
+			ssh.PublicKeys(signer),
+		},
+		HostKeyCallback: func(hostname string, remote net.Addr, key ssh.PublicKey) error {
+			// use OpenSSH's known_hosts file if you care about host validation
+			return nil
+		},
 	}
 	conn, err := ssh.Dial("tcp", net.JoinHostPort(ip, sshPort), config)
 	if err != nil {
@@ -89,59 +97,15 @@ func executeCleanupOnRemote(sshUser string, sshPort string, sshKeyFile string, i
 	var stdoutBuf bytes.Buffer
 	session.Stdout = &stdoutBuf
 	writer.Printf("cleaning up the nodes on IP: " + ip)
+	if ip == "FrontendIps" {
 	err = session.Run("sudo systemctl stop chef-automate;sudo rm -rf /hab")
+	} else {
+		err = session.Run("sudo systemctl stop hab-sup;sudo rm -rf /hab")
+	}
 	if err != nil {
 		writer.Errorf("Run failed:%v", err)
 	} else {
 		writer.Success("Destroy successful...\n")
 	}
 	defer session.Close()
-	writer.Printf(">%s", stdoutBuf)
-	// script :=fmt.Sprintf(FSCRIPT)
-	// command := exec.Command(script)
-	// err = command.Run()
-	// if err != nil {
-	// 	writer.Fail(err.Error())
-	// 	//return err
-	// }
-}
-// err = session.Run("systemctl stop chef-automate;
-// rm -rf /var/automate-ha;
-// rm -rf /hab")
-
-func executeCleanupOnBackendRemote(sshUser string, sshPort string, sshKeyFile string, ip string, remoteType string) {
-
-	pemBytes, err := ioutil.ReadFile(sshKeyFile)
-	if err != nil {
-		writer.Errorf("Unable to read private key: %v", err)
-	}
-	signer, err := ssh.ParsePrivateKey(pemBytes)
-	if err != nil {
-		writer.Errorf("Parsing key failed: %v", err)
-	}
-	config := &ssh.ClientConfig{
-		User:            sshUser,
-		Auth:            []ssh.AuthMethod{ssh.PublicKeys(signer)},
-		HostKeyCallback: ssh.InsecureIgnoreHostKey(),
-	}
-	conn, err := ssh.Dial("tcp", net.JoinHostPort(ip, sshPort), config)
-	if err != nil {
-		writer.Errorf("dial failed:%v", err)
-	}
-	defer conn.Close()
-	session, err := conn.NewSession()
-	if err != nil {
-		writer.Errorf("session failed:%v", err)
-	}
-	var stdoutBuf bytes.Buffer
-	session.Stdout = &stdoutBuf
-	writer.Printf("cleaning up the nodes on IP: " + ip)
-	err = session.Run("sudo systemctl stop hab-sup;sudo rm -rf /hab")
-	if err != nil {
-		writer.Errorf("Run failed:%v", err)
-	} else {
-		writer.Success("Destroy successful...\n")
-	}
-	defer session.Close()
-	writer.Printf(">%s", stdoutBuf)
 }
